directories refactoring
