+++ src/personal_path/models/course-preview-info.js
