+++ src/personal_path/views/course-container.js
