+++ src/personal_path/views/course-preview.js
