+++ src/personal_path/views/course-view.js
